+++ spec/mtl_spec.rb
@@ -12,6 +12,10 @@ RSpec.describe Mtl do
     JSON.parse(File.read('node_modules/material-design-icons/package.json'))['version']
   end
 
+  let(:lodash_version) do
+    JSON.parse(File.read('node_modules/lodash/package.json'))['version']
+  end
+
   it 'has a version number' do
     expect(subject::VERSION).not_to be nil
   end
@@ -22,5 +26,9 @@ RSpec.describe Mtl do
 
   it 'has a ICONS_VERSION' do
     expect(subject::ICONS_VERSION).to eq icons_version
+  end
+
+  it 'has a LODASH_VERSION' do
+    expect(subject::LODASH_VERSION).to eq lodash_version
   end
 end